requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
google-auth==2.36.0
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.1
pandas==2.2.3
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==14.0
